Revert "+ lib/football-data module and test"
This reverts commit 26fc803395b1c500172e22c9caae3bedd1dc3712.
